Redo some arena stub concepts

+++ 3dDreams/vulkan_surface.c
@@ -1,11 +1,11 @@
 #include "vulkan.h"
 #include "common.h"
 
-static bool vulkan_window_surface_create(vulkan_context* context, const hw_window* window, const char** extension_names, u32 extension_count)
+static bool vulkan_window_surface_create(vulkan_context* context, const hw_window* window, const char** extension_names, usize extension_count)
 {
    bool isWin32Surface = false;
 
-   for(size_t i = 0; i < extension_count; ++i)
+   for(usize i = 0; i < extension_count; ++i)
       if(strcmp(extension_names[i], VK_KHR_WIN32_SURFACE_EXTENSION_NAME) == 0) {
          isWin32Surface = true;
          break;
